**IMPORTANT UPDATE** : update a compiler to 'gcc Built by MinGW-W64 project 4.9.1'. Accordingly, now 64bit architecture is the LLP64(previously, is the LP64). FIXED: Directory whose name contains '.' causes failed to load a proper icon.

+++ src/ib_platform_win.h
@@ -13,7 +13,7 @@
 #include <lm.h>
 
 namespace ib{
-  typedef unsigned long thread;
+  typedef intptr_t thread;
   typedef void (*threadfunc) (void *);
   typedef CRITICAL_SECTION mutex;
   typedef HANDLE condition;
